requirements.txt
django>=5.0.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
psycopg2-binary
Pillow
gunicorn
whitenoise
dj-database-url
